+++ just_audio/darwin/Classes/JustAudioPlugin.m
@@ -27,12 +27,13 @@
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
     if ([@"init" isEqualToString:call.method]) {
         NSDictionary *request = (NSDictionary *)call.arguments;
-        NSString *playerId = request[@"id"];
+        NSString *playerId = (NSString *)request[@"id"];
+        NSDictionary *loadConfiguration = (NSDictionary *)request[@"audioLoadConfiguration"];
         if ([_players objectForKey:playerId] != nil) {
             FlutterError *flutterError = [FlutterError errorWithCode:@"error" message:@"Platform player already exists" details:nil];
             result(flutterError);
         } else {
-            AudioPlayer* player = [[AudioPlayer alloc] initWithRegistrar:_registrar playerId:playerId];
+            AudioPlayer* player = [[AudioPlayer alloc] initWithRegistrar:_registrar playerId:playerId loadConfiguration:loadConfiguration];
             [_players setValue:player forKey:playerId];
             result(nil);
         }
@@ -41,6 +42,12 @@
         NSString *playerId = request[@"id"];
         [_players[playerId] dispose];
         [_players setValue:nil forKey:playerId];
+        result(@{});
+    } else if ([@"disposeAllPlayers" isEqualToString:call.method]) {
+        for (NSString *playerId in _players) {
+            [_players[playerId] dispose];
+        }
+        [_players removeAllObjects];
         result(@{});
     } else {
         result(FlutterMethodNotImplemented);