+++ just_audio/test/just_audio_test.dart
@@ -4,6 +4,7 @@ import 'dart:io';
 import 'dart:math';
 import 'dart:typed_data';
 
+import 'package:audio_session/audio_session.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:just_audio/just_audio.dart';
@@ -22,7 +23,7 @@ void main() {
 void runTests() {
   final mock = MockJustAudio();
   JustAudioPlatform.instance = mock;
-  final audioSessionChannel = MethodChannel('com.ryanheise.audio_session');
+  const audioSessionChannel = MethodChannel('com.ryanheise.audio_session');
 
   void expectDuration(Duration a, Duration b, {int epsilon = 200}) {
     expect((a - b).inMilliseconds.abs(), lessThanOrEqualTo(epsilon));
@@ -96,8 +97,9 @@ void runTests() {
     expect(player.loopMode, equals(LoopMode.one));
     await player.setShuffleModeEnabled(true);
     expect(player.shuffleModeEnabled, equals(true));
-    await player.seek(Duration(seconds: 1));
-    expectState(player: player, playing: false, position: Duration(seconds: 1));
+    await player.seek(const Duration(seconds: 1));
+    expectState(
+        player: player, playing: false, position: const Duration(seconds: 1));
     final playlist = ConcatenatingAudioSource(children: []);
     await player.setAudioSource(playlist, preload: false);
     await playlist.addAll([
@@ -188,26 +190,26 @@ void runTests() {
       position: point1,
       processingState: ProcessingState.ready,
     );
-    await Future<dynamic>.delayed(Duration(milliseconds: 100));
+    await Future<dynamic>.delayed(const Duration(milliseconds: 100));
     expectState(player: player, playing: true);
-    await Future<dynamic>.delayed(Duration(seconds: 1));
+    await Future<dynamic>.delayed(const Duration(seconds: 1));
     expectState(
       player: player,
-      position: point1 + Duration(seconds: 1),
+      position: point1 + const Duration(seconds: 1),
       processingState: ProcessingState.ready,
       playing: true,
     );
-    await player.seek(duration - Duration(seconds: 3));
+    await player.seek(duration - const Duration(seconds: 3));
     expectState(
       player: player,
-      position: duration - Duration(seconds: 3),
+      position: duration - const Duration(seconds: 3),
       processingState: ProcessingState.ready,
       playing: true,
     );
     await player.pause();
     expectState(
       player: player,
-      position: duration - Duration(seconds: 3),
+      position: duration - const Duration(seconds: 3),
       processingState: ProcessingState.ready,
       playing: false,
     );
@@ -216,12 +218,12 @@ void runTests() {
     final playFuture = player.play();
     expectState(
       player: player,
-      position: duration - Duration(seconds: 3),
+      position: duration - const Duration(seconds: 3),
       processingState: ProcessingState.ready,
     );
     expectState(player: player, playing: true);
     await playFuture;
-    expectDuration(stopwatch.elapsed, Duration(seconds: 3));
+    expectDuration(stopwatch.elapsed, const Duration(seconds: 3));
     expectState(
       player: player,
       position: duration,
@@ -234,10 +236,10 @@ void runTests() {
   test('speed', () async {
     final player = AudioPlayer();
     /*final duration =*/ await player.setUrl('https://foo.foo/foo.mp3');
-    final period1 = Duration(seconds: 2);
-    final period2 = Duration(seconds: 2);
-    final speed1 = 0.75;
-    final speed2 = 1.5;
+    const period1 = Duration(seconds: 2);
+    const period2 = Duration(seconds: 2);
+    const speed1 = 0.75;
+    const speed2 = 1.5;
     final position1 = period1 * speed1;
     final position2 = position1 + period2 * speed2;
     expectState(player: player, position: Duration.zero);
@@ -251,14 +253,69 @@ void runTests() {
     await player.dispose();
   });
 
+  test('skipSilence', () async {
+    final player = AudioPlayer();
+    expect(player.skipSilenceEnabled, equals(false));
+    await player.setSkipSilenceEnabled(true);
+    expect(player.skipSilenceEnabled, equals(true));
+    await player.setSkipSilenceEnabled(false);
+    expect(player.skipSilenceEnabled, equals(false));
+    await player.dispose();
+  });
+
+  test('pitch', () async {
+    final player = AudioPlayer();
+    expect(player.pitch, equals(1.0));
+    await player.setPitch(1.5);
+    expect(player.pitch, equals(1.5));
+    await player.dispose();
+  });
+
+  test('setAutomaticallyWaitsToMinimizeStalling', () async {
+    final player = AudioPlayer();
+    expect(player.automaticallyWaitsToMinimizeStalling, equals(true));
+    await player.setAutomaticallyWaitsToMinimizeStalling(false);
+    expect(player.automaticallyWaitsToMinimizeStalling, equals(false));
+    await player.setAutomaticallyWaitsToMinimizeStalling(true);
+    expect(player.automaticallyWaitsToMinimizeStalling, equals(true));
+    await player.dispose();
+  });
+
+  test('setCanUseNetworkResourcesForLiveStreamingWhilePaused', () async {
+    final player = AudioPlayer();
+    expect(player.canUseNetworkResourcesForLiveStreamingWhilePaused,
+        equals(false));
+    await player.setCanUseNetworkResourcesForLiveStreamingWhilePaused(true);
+    expect(
+        player.canUseNetworkResourcesForLiveStreamingWhilePaused, equals(true));
+    await player.setCanUseNetworkResourcesForLiveStreamingWhilePaused(false);
+    expect(player.canUseNetworkResourcesForLiveStreamingWhilePaused,
+        equals(false));
+    await player.dispose();
+  });
+
+  test('setPreferredPeakBitRate', () async {
+    final player = AudioPlayer();
+    expect(player.preferredPeakBitRate, equals(0.0));
+    await player.setPreferredPeakBitRate(1000.0);
+    expect(player.preferredPeakBitRate, equals(1000.0));
+    await player.dispose();
+  });
+
+  test('setAndroidAudioAttributes', () async {
+    final player = AudioPlayer();
+    await player.setAndroidAudioAttributes(const AndroidAudioAttributes());
+    await player.dispose();
+  });
+
   test('positionStream', () async {
     final player = AudioPlayer();
     /*final duration =*/ await player.setUrl('https://foo.foo/foo.mp3');
-    final period = Duration(seconds: 3);
-    final position1 = period;
+    const period = Duration(seconds: 3);
+    const position1 = period;
     final position2 = position1 + period;
-    final speed1 = 0.75;
-    final speed2 = 1.5;
+    const speed1 = 0.75;
+    const speed2 = 1.5;
     final stepDuration = period ~/ 5;
     var target = stepDuration;
     player.setSpeed(speed1);
@@ -389,7 +446,7 @@ void runTests() {
         expect(response.statusCode, equals(HttpStatus.partialContent));
       } else if (start != null) {
         expect(response.headers.value(HttpHeaders.contentRangeHeader),
-            'bytes $start-/${byteRangeData.length}');
+            'bytes $start-${byteRangeData.length - 1}/${byteRangeData.length}');
         expect(response.statusCode, equals(HttpStatus.partialContent));
       } else {
         expect(response.headers.value(HttpHeaders.contentRangeHeader), null);
@@ -412,8 +469,8 @@ void runTests() {
       LoopingAudioSource(
         count: 2,
         child: ClippingAudioSource(
-          start: Duration(seconds: 60),
-          end: Duration(seconds: 65),
+          start: const Duration(seconds: 60),
+          end: const Duration(seconds: 65),
           child: AudioSource.uri(Uri.parse("https://foo.foo/foo.mp3")),
           tag: 'a',
         ),
@@ -469,6 +526,48 @@ void runTests() {
     await source2.clear();
     expect(source2.sequence.map((s) => s.tag as String?), equals(<String>[]));
     await player.dispose();
+  });
+
+  test('idle-sequence', () async {
+    final source = ConcatenatingAudioSource(children: []);
+    final player = AudioPlayer();
+    await player.setAudioSource(source);
+    await player.stop();
+    expect(source.sequence.length, equals(0));
+    await source.add(AudioSource.uri(Uri.parse('https://b.b/b.mp3'), tag: 'b'));
+    await source.insert(
+        0, AudioSource.uri(Uri.parse('https://a.a/a.mp3'), tag: 'a'));
+    await source.insert(
+        2, AudioSource.uri(Uri.parse('https://c.c/c.mp3'), tag: 'c'));
+    await source.addAll([
+      AudioSource.uri(Uri.parse('https://d.d/d.mp3'), tag: 'd'),
+      AudioSource.uri(Uri.parse('https://e.e/e.mp3'), tag: 'e'),
+    ]);
+    await source.insertAll(3, [
+      AudioSource.uri(Uri.parse('https://e.e/e.mp3'), tag: 'e'),
+      AudioSource.uri(Uri.parse('https://f.f/f.mp3'), tag: 'f'),
+    ]);
+    expect(source.sequence.map((s) => s.tag as String?),
+        equals(['a', 'b', 'c', 'e', 'f', 'd', 'e']));
+    await source.removeAt(0);
+    expect(source.sequence.map((s) => s.tag as String?),
+        equals(['b', 'c', 'e', 'f', 'd', 'e']));
+    await source.move(3, 2);
+    expect(source.sequence.map((s) => s.tag as String?),
+        equals(['b', 'c', 'f', 'e', 'd', 'e']));
+    await source.move(2, 3);
+    expect(source.sequence.map((s) => s.tag as String?),
+        equals(['b', 'c', 'e', 'f', 'd', 'e']));
+    await source.removeRange(0, 2);
+    expect(source.sequence.map((s) => s.tag as String?),
+        equals(['e', 'f', 'd', 'e']));
+    await source.removeAt(3);
+    expect(
+        source.sequence.map((s) => s.tag as String?), equals(['e', 'f', 'd']));
+    await source.removeRange(1, 3);
+    expect(source.sequence.map((s) => s.tag as String?), equals(['e']));
+    await source.clear();
+    expect(source.sequence.map((s) => s.tag as String?), equals(<String>[]));
   });
 
   test('sequence-state', () async {
@@ -547,13 +646,14 @@ void runTests() {
     final player = AudioPlayer();
     final duration1 = await player.setUrl('https://bar.bar/foo.mp3');
     expectDuration(duration1!, audioSourceDuration);
-    final duration2 = await player.setClip(start: Duration(seconds: 10));
-    expectDuration(duration2!, audioSourceDuration - Duration(seconds: 10));
-    final duration3 = await player.setClip(end: Duration(seconds: 15));
-    expectDuration(duration3!, Duration(seconds: 15));
+    final duration2 = await player.setClip(start: const Duration(seconds: 10));
+    expectDuration(
+        duration2!, audioSourceDuration - const Duration(seconds: 10));
+    final duration3 = await player.setClip(end: const Duration(seconds: 15));
+    expectDuration(duration3!, const Duration(seconds: 15));
     final duration4 = await player.setClip(
-        start: Duration(seconds: 10), end: Duration(seconds: 21));
-    expectDuration(duration4!, Duration(seconds: 11));
+        start: const Duration(seconds: 10), end: const Duration(seconds: 21));
+    expectDuration(duration4!, const Duration(seconds: 11));
     final duration5 = await player.setClip();
     expectDuration(duration5!, audioSourceDuration);
     await player.dispose();
@@ -618,8 +718,8 @@ void runTests() {
             LoopingAudioSource(
               count: 2,
               child: ClippingAudioSource(
-                start: Duration(seconds: 60),
-                end: Duration(seconds: 65),
+                start: const Duration(seconds: 60),
+                end: const Duration(seconds: 65),
                 child: AudioSource.uri(Uri.parse("https://foo.foo/foo.mp3")),
                 tag: 'a',
               ),
@@ -721,21 +821,21 @@ void runTests() {
     expect(player.processingState, ProcessingState.idle);
     await player.load();
     expect(player.processingState, ProcessingState.ready);
-    await player.seek(Duration(seconds: 5), index: 1);
+    await player.seek(const Duration(seconds: 5), index: 1);
     await player.setVolume(0.5);
     await player.setSpeed(0.7);
     await player.setShuffleModeEnabled(true);
     await player.setLoopMode(LoopMode.one);
     await player.stop();
     expect(player.processingState, ProcessingState.idle);
-    expect(player.position, Duration(seconds: 5));
+    expect(player.position, const Duration(seconds: 5));
     expect(player.volume, 0.5);
     expect(player.speed, 0.7);
     expect(player.shuffleModeEnabled, true);
     expect(player.loopMode, LoopMode.one);
     await player.load();
     expect(player.processingState, ProcessingState.ready);
-    expect(player.position, Duration(seconds: 5));
+    expect(player.position, const Duration(seconds: 5));
     expect(player.volume, 0.5);
     expect(player.speed, 0.7);
     expect(player.shuffleModeEnabled, true);
@@ -756,8 +856,8 @@ void runTests() {
       expect(player.processingState, equals(ProcessingState.ready));
       expect(player.playing, equals(true));
       expectDuration(player.position, Duration.zero);
-      await Future<dynamic>.delayed(Duration(seconds: 1));
-      expectDuration(player.position, Duration(seconds: 1));
+      await Future<dynamic>.delayed(const Duration(seconds: 1));
+      expectDuration(player.position, const Duration(seconds: 1));
       await player.dispose();
     }
   });
@@ -775,8 +875,8 @@ void runTests() {
       expect(player.processingState, equals(ProcessingState.ready));
       expect(player.playing, equals(true));
       expectDuration(player.position, Duration.zero);
-      await Future<dynamic>.delayed(Duration(seconds: 1));
-      expectDuration(player.position, Duration(seconds: 1));
+      await Future<dynamic>.delayed(const Duration(seconds: 1));
+      expectDuration(player.position, const Duration(seconds: 1));
       await player.dispose();
     }
   });
@@ -793,8 +893,8 @@ void runTests() {
     expect(player.processingState, equals(ProcessingState.ready));
     expect(player.playing, equals(true));
     expectDuration(player.position, Duration.zero);
-    await Future<dynamic>.delayed(Duration(seconds: 1));
-    expectDuration(player.position, Duration(seconds: 1));
+    await Future<dynamic>.delayed(const Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 1));
     await player.dispose();
   });
 
@@ -895,15 +995,15 @@ void runTests() {
     await player.setUrl('https://bar.bar/foo.mp3');
     expect(player.processingState, equals(ProcessingState.ready));
     expect(player.playing, equals(true));
-    expectDuration(player.position, Duration(seconds: 0));
-    await Future<dynamic>.delayed(Duration(seconds: 1));
-    expectDuration(player.position, Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 0));
+    await Future<dynamic>.delayed(const Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 1));
     await player.setUrl('https://bar.bar/bar.mp3');
     expect(player.processingState, equals(ProcessingState.ready));
     expect(player.playing, equals(true));
-    expectDuration(player.position, Duration(seconds: 0));
-    await Future<dynamic>.delayed(Duration(seconds: 1));
-    expectDuration(player.position, Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 0));
+    await Future<dynamic>.delayed(const Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 1));
     await player.dispose();
   });
 
@@ -913,9 +1013,9 @@ void runTests() {
     await player.setUrl('https://bar.bar/foo.mp3');
     expect(player.processingState, equals(ProcessingState.ready));
     expect(player.playing, equals(true));
-    expectDuration(player.position, Duration(seconds: 0));
-    await Future<dynamic>.delayed(Duration(seconds: 1));
-    expectDuration(player.position, Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 0));
+    await Future<dynamic>.delayed(const Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 1));
     player.pause();
     expect(player.playing, equals(false));
     await player.setUrl('https://bar.bar/bar.mp3', preload: false);
@@ -925,12 +1025,12 @@ void runTests() {
     await player.load();
     expect(player.processingState, equals(ProcessingState.ready));
     expect(player.playing, equals(false));
-    expectDuration(player.position, Duration(seconds: 0));
+    expectDuration(player.position, const Duration(seconds: 0));
     player.play();
     expect(player.playing, equals(true));
-    expectDuration(player.position, Duration(seconds: 0));
-    await Future<dynamic>.delayed(Duration(seconds: 1));
-    expectDuration(player.position, Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 0));
+    await Future<dynamic>.delayed(const Duration(seconds: 1));
+    expectDuration(player.position, const Duration(seconds: 1));
     await player.dispose();
   });
 
@@ -941,11 +1041,306 @@ void runTests() {
     await player.setUrl('https://bar.bar/foo.mp3');
     await player.dispose();
   });
+
+  test('play-pause', () async {
+    final player = AudioPlayer();
+    await player.setUrl('https://bar.bar/foo.mp3');
+    player.play();
+    await player.pause();
+    await Future<void>.delayed(const Duration(milliseconds: 200));
+    expect(mock.mostRecentPlayer?._playing, false);
+    await player.dispose();
+  });
+
+  test('play-stop', () async {
+    final player = AudioPlayer();
+    await player.setUrl('https://bar.bar/foo.mp3');
+    player.play();
+    await player.stop();
+    await Future<void>.delayed(const Duration(milliseconds: 200));
+    expect(mock.mostRecentPlayer?._playing, false);
+    await player.dispose();
+  });
+
+  test('positionStream emissions: seek while paused', () async {
+    final player = AudioPlayer();
+    await player.setUrl('https://bar.bar/foo.mp3');
+    expectState(
+      player: player,
+      position: Duration.zero,
+      processingState: ProcessingState.ready,
+      playing: false,
+    );
+    var completer = Completer<dynamic>();
+    late StreamSubscription subscription;
+    subscription = player.positionStream.listen((position) {
+      expectDuration(position, Duration.zero);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    const duration1 = Duration(seconds: 1);
+    const duration2 = Duration(milliseconds: 600);
+    const duration3 = Duration(milliseconds: 750);
+
+    await player.seek(duration1);
+    expectState(
+      player: player,
+      position: duration1,
+      processingState: ProcessingState.ready,
+      playing: false,
+    );
+    completer = Completer<dynamic>();
+    subscription = player.positionStream.listen((position) {
+      expectDuration(position, duration1);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    player.play();
+    await Future<dynamic>.delayed(duration2);
+    expectState(
+      player: player,
+      position: duration1 + duration2,
+      processingState: ProcessingState.ready,
+      playing: true,
+    );
+    await player.pause();
+    completer = Completer<dynamic>();
+    subscription = player.positionStream.listen((position) {
+      expectDuration(position, duration1 + duration2);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    await player.seek(duration1 + duration2 + duration3);
+    expectState(
+      player: player,
+      position: duration1 + duration2 + duration3,
+      processingState: ProcessingState.ready,
+      playing: false,
+    );
+    completer = Completer<dynamic>();
+    subscription = player.positionStream.listen((position) {
+      expectDuration(position, duration1 + duration2 + duration3);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    await player.dispose();
+  });
+
+  test('positionStream emissions: switch audio sources', () async {
+    final player = AudioPlayer();
+    final playlist = ConcatenatingAudioSource(
+      children: [
+        AudioSource.uri(Uri.parse('https://bar.bar/foo.mp3')),
+        AudioSource.uri(Uri.parse('https://bar.bar/bar.mp3')),
+      ],
+    );
+    await player.setAudioSource(playlist);
+    expectState(
+      player: player,
+      position: Duration.zero,
+      processingState: ProcessingState.ready,
+      playing: false,
+    );
+    expect(player.currentIndex, 0);
+    var completer = Completer<dynamic>();
+    late StreamSubscription subscription;
+    subscription = player.positionStream.listen((position) {
+      expectDuration(position, Duration.zero);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    const duration1 = Duration(seconds: 1);
+    const duration2 = Duration(seconds: 600);
+
+    await player.seek(duration1);
+    expect(player.currentIndex, 0);
+    completer = Completer<dynamic>();
+    subscription = player.positionStream.listen((position) {
+      expectDuration(position, duration1);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    await player.seekToNext();
+    expect(player.currentIndex, 1);
+    completer = Completer<dynamic>();
+    subscription = player.positionStream.listen((position) {
+      expectDuration(position, Duration.zero);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    await player.seek(duration1, index: 1);
+    expect(player.currentIndex, 1);
+    await player.seekToNext();
+    // There is no next
+    expect(player.currentIndex, 1);
+    completer = Completer<dynamic>();
+    subscription = player.positionStream.listen((position) {
+      // No position change because there is no next
+      expectDuration(position, duration1);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    // Switch to index 0 and seek position at the same time
+    await player.seek(duration2, index: 0);
+    expect(player.currentIndex, 0);
+    completer = Completer<dynamic>();
+    subscription = player.positionStream.listen((position) {
+      expectDuration(position, duration2);
+      subscription.cancel();
+      completer.complete();
+    });
+    await completer.future;
+
+    await player.dispose();
+  });
+
+  test('positionDiscontinuity', () async {
+    final player = AudioPlayer();
+    final discontinuityEvents = <PositionDiscontinuity>[];
+    final subscription =
+        player.positionDiscontinuityStream.listen(discontinuityEvents.add);
+    final playlist = ConcatenatingAudioSource(children: [
+      AudioSource.uri(
+        Uri.parse("https://bar.bar/bar.mp3"),
+        tag: 'a',
+      ),
+      AudioSource.uri(
+        Uri.parse("https://baz.baz/baz.mp3"),
+        tag: 'b',
+      ),
+    ]);
+    await player.setAudioSource(playlist);
+    expect(player.currentIndex, equals(0));
+    expect(discontinuityEvents.length, equals(0));
+    player.play();
+    await Future<void>.delayed(const Duration(milliseconds: 500));
+    expect(discontinuityEvents.length, equals(0));
+    for (var mockPlayer in mock._players.values) {
+      await mockPlayer._autoAdvance();
+    }
+    expect(player.currentIndex, equals(1));
+    expect(
+        discontinuityEvents.length == 1 &&
+            discontinuityEvents.first.reason ==
+                PositionDiscontinuityReason.autoAdvance,
+        equals(true));
+    discontinuityEvents.clear();
+    await player.seek(Duration.zero, index: 0);
+    expect(player.currentIndex, equals(0));
+    expect(
+        discontinuityEvents.length == 1 &&
+            discontinuityEvents.first.reason ==
+                PositionDiscontinuityReason.seek,
+        equals(true));
+    discontinuityEvents.clear();
+    // Test loop one
+    await player.setLoopMode(LoopMode.one);
+    await player.seek(const Duration(seconds: 119));
+    discontinuityEvents.clear();
+    for (var mockPlayer in mock._players.values) {
+      await mockPlayer._autoAdvance();
+    }
+    expect(player.currentIndex, equals(0));
+    expect(
+        discontinuityEvents.length == 1 &&
+            discontinuityEvents.first.reason ==
+                PositionDiscontinuityReason.autoAdvance,
+        equals(true));
+    discontinuityEvents.clear();
+
+    await player.dispose();
+    subscription.cancel();
+  });
+
+  test('loadConfiguration', () async {
+    final audioLoadConfiguration = AudioLoadConfiguration(
+      darwinLoadControl: DarwinLoadControl(),
+      androidLoadControl: AndroidLoadControl(),
+      androidLivePlaybackSpeedControl: AndroidLivePlaybackSpeedControl(),
+    );
+    final player = AudioPlayer(
+      audioLoadConfiguration: audioLoadConfiguration,
+    );
+    await player.setUrl('https://foo.foo/foo.mp3');
+    final platformPlayer = mock.mostRecentPlayer!;
+    expect(
+        platformPlayer.audioLoadConfiguration?.darwinLoadControl
+            ?.automaticallyWaitsToMinimizeStalling,
+        equals(audioLoadConfiguration
+            .darwinLoadControl?.automaticallyWaitsToMinimizeStalling));
+    // TODO: check other fields.
+    await player.dispose();
+  });
+
+  test('AndroidLoudnessEnhancer', () async {
+    final loudnessEnhancer = AndroidLoudnessEnhancer();
+    final player = AudioPlayer(
+      audioPipeline: AudioPipeline(androidAudioEffects: [loudnessEnhancer]),
+    );
+    expect(loudnessEnhancer.targetGain, equals(0.0));
+    expect(await loudnessEnhancer.targetGainStream.first, equals(0.0));
+    await player.setUrl('https://foo.foo/foo.mp3');
+    expect(loudnessEnhancer.targetGain, equals(0.0));
+    expect(await loudnessEnhancer.targetGainStream.first, equals(0.0));
+    await loudnessEnhancer.setTargetGain(1.5);
+    expect(loudnessEnhancer.targetGain, equals(1.5));
+    expect(await loudnessEnhancer.targetGainStream.first, equals(1.5));
+    expect(loudnessEnhancer.enabled, equals(false));
+    expect(await loudnessEnhancer.enabledStream.first, equals(false));
+    await loudnessEnhancer.setEnabled(true);
+    expect(loudnessEnhancer.enabled, equals(true));
+    expect(await loudnessEnhancer.enabledStream.first, equals(true));
+  });
+
+  test('AndroidEqualizer', () async {
+    final equalizer = AndroidEqualizer();
+    final player = AudioPlayer(
+      audioPipeline: AudioPipeline(androidAudioEffects: [equalizer]),
+    );
+    expect(equalizer.enabled, equals(false));
+    expect(await equalizer.enabledStream.first, equals(false));
+    await player.setUrl('https://foo.foo/foo.mp3');
+    expect(equalizer.enabled, equals(false));
+    expect(await equalizer.enabledStream.first, equals(false));
+    await equalizer.setEnabled(true);
+    expect(equalizer.enabled, equals(true));
+    expect(await equalizer.enabledStream.first, equals(true));
+    final parameters = await equalizer.parameters;
+    expect(parameters.minDecibels, equals(0.0));
+    expect(parameters.maxDecibels, equals(10.0));
+    final bands = parameters.bands;
+    expect(bands.length, equals(5));
+    for (var i = 0; i < 5; i++) {
+      final band = bands[i];
+      expect(band.index, equals(i));
+      expect(band.lowerFrequency, equals(i * 1000));
+      expect(band.upperFrequency, equals((i + 1) * 1000));
+      expect(band.centerFrequency, equals((i + 0.5) * 1000));
+      expect(band.gain, equals(i * 0.1));
+    }
+  });
 }
 
 class MockJustAudio extends Mock
     with MockPlatformInterfaceMixin
     implements JustAudioPlatform {
+  MockAudioPlayer? mostRecentPlayer;
   final _players = <String, MockAudioPlayer>{};
 
   @override
@@ -955,8 +1350,9 @@ class MockJustAudio extends Mock
           code: "error",
           message: "Platform player ${request.id} already exists");
     }
-    final player = MockAudioPlayer(request.id);
+    final player = MockAudioPlayer(request);
     _players[request.id] = player;
+    mostRecentPlayer = player;
     return player;
   }
 
@@ -966,6 +1362,16 @@ class MockJustAudio extends Mock
     _players[request.id]!.dispose(DisposeRequest());
     _players.remove(request.id);
     return DisposePlayerResponse();
+  }
+
+  @override
+  Future<DisposeAllPlayersResponse> disposeAllPlayers(
+      DisposeAllPlayersRequest request) async {
+    for (var player in _players.values) {
+      player.dispose(DisposeRequest());
+    }
+    _players.clear();
+    return DisposeAllPlayersResponse();
   }
 }
 
@@ -1001,9 +1407,9 @@ final icyMetadataMessage = IcyMetadataMessage(
   ),
 );
 
-class MockAudioPlayer implements AudioPlayerPlatform {
-  final String _id;
+class MockAudioPlayer extends AudioPlayerPlatform {
   final eventController = StreamController<PlaybackEventMessage>();
+  final AudioLoadConfigurationMessage? audioLoadConfiguration;
   AudioSourceMessage? _audioSource;
   ProcessingStateMessage _processingState = ProcessingStateMessage.idle;
   Duration _updatePosition = Duration.zero;
@@ -1015,11 +1421,15 @@ class MockAudioPlayer implements AudioPlayerPlatform {
   var _speed = 1.0;
   Completer<dynamic>? _playCompleter;
   Timer? _playTimer;
+  LoopModeMessage _loopMode = LoopModeMessage.off;
 
-  MockAudioPlayer(String id) : _id = id;
+  MockAudioPlayer(InitRequest request)
+      : audioLoadConfiguration = request.audioLoadConfiguration,
+        super(request.id);
 
   @override
-  String get id => _id;
+  Stream<PlayerDataMessage> get playerDataMessageStream =>
+      StreamController<PlayerDataMessage>().stream;
 
   @override
   Stream<PlaybackEventMessage> get playbackEventMessageStream =>
@@ -1028,6 +1438,8 @@ class MockAudioPlayer implements AudioPlayerPlatform {
   @override
   Future<LoadResponse> load(LoadRequest request) async {
     final audioSource = request.audioSourceMessage;
+    _processingState = ProcessingStateMessage.loading;
+    _broadcastPlaybackEvent();
     if (audioSource is UriAudioSourceMessage) {
       if (audioSource.uri.contains('abort')) {
         throw PlatformException(code: 'abort', message: 'Failed to load URL');
@@ -1048,7 +1460,7 @@ class MockAudioPlayer implements AudioPlayerPlatform {
     _audioSource = audioSource;
     _index = request.initialIndex ?? 0;
     // Simulate loading time.
-    await Future<dynamic>.delayed(Duration(milliseconds: 100));
+    await Future<dynamic>.delayed(const Duration(milliseconds: 100));
     _setPosition(request.initialPosition ?? Duration.zero);
     _processingState = ProcessingStateMessage.ready;
     _broadcastPlaybackEvent();
@@ -1066,7 +1478,6 @@ class MockAudioPlayer implements AudioPlayerPlatform {
       _startTimer();
     }
     _playCompleter = Completer<dynamic>();
-    _broadcastPlaybackEvent();
     await _playCompleter!.future;
     return PlayResponse();
   }
@@ -1094,9 +1505,17 @@ class MockAudioPlayer implements AudioPlayerPlatform {
   @override
   Future<SeekResponse> seek(SeekRequest request) async {
     _setPosition(request.position ?? Duration.zero);
-    _index = request.index;
+    _index = request.index ?? 0;
     _broadcastPlaybackEvent();
     return SeekResponse();
+  }
+
+  Future<void> _autoAdvance() async {
+    _setPosition(Duration.zero);
+    if (_loopMode == LoopModeMessage.off) {
+      _index = _index! + 1;
+    }
+    _broadcastPlaybackEvent();
   }
 
   @override
@@ -1121,6 +1540,7 @@ class MockAudioPlayer implements AudioPlayerPlatform {
 
   @override
   Future<SetLoopModeResponse> setLoopMode(SetLoopModeRequest request) async {
+    _loopMode = request.loopMode;
     return SetLoopModeResponse();
   }
 
@@ -1141,6 +1561,17 @@ class MockAudioPlayer implements AudioPlayerPlatform {
     _speed = request.speed;
     _setPosition(_position);
     return SetSpeedResponse();
+  }
+
+  @override
+  Future<SetPitchResponse> setPitch(SetPitchRequest request) async {
+    return SetPitchResponse();
+  }
+
+  @override
+  Future<SetSkipSilenceResponse> setSkipSilence(
+      SetSkipSilenceRequest request) async {
+    return SetSkipSilenceResponse();
   }
 
   @override
@@ -1223,12 +1654,66 @@ class MockAudioPlayer implements AudioPlayerPlatform {
     _updatePosition = position;
     _updateTime = DateTime.now();
   }
+
+  @override
+  Future<SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse>
+      setCanUseNetworkResourcesForLiveStreamingWhilePaused(
+          SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest
+              request) async {
+    return SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse();
+  }
+
+  @override
+  Future<SetPreferredPeakBitRateResponse> setPreferredPeakBitRate(
+      SetPreferredPeakBitRateRequest request) async {
+    return SetPreferredPeakBitRateResponse();
+  }
+
+  @override
+  Future<AudioEffectSetEnabledResponse> audioEffectSetEnabled(
+      AudioEffectSetEnabledRequest request) async {
+    return AudioEffectSetEnabledResponse();
+  }
+
+  @override
+  Future<AndroidLoudnessEnhancerSetTargetGainResponse>
+      androidLoudnessEnhancerSetTargetGain(
+          AndroidLoudnessEnhancerSetTargetGainRequest request) async {
+    return AndroidLoudnessEnhancerSetTargetGainResponse();
+  }
+
+  @override
+  Future<AndroidEqualizerGetParametersResponse> androidEqualizerGetParameters(
+      AndroidEqualizerGetParametersRequest request) async {
+    return AndroidEqualizerGetParametersResponse(
+      parameters: AndroidEqualizerParametersMessage(
+        minDecibels: 0.0,
+        maxDecibels: 10.0,
+        bands: [
+          for (var i = 0; i < 5; i++)
+            AndroidEqualizerBandMessage(
+              index: i,
+              lowerFrequency: i * 1000,
+              upperFrequency: (i + 1) * 1000,
+              centerFrequency: (i + 0.5) * 1000,
+              gain: i * 0.1,
+            ),
+        ],
+      ),
+    );
+  }
+
+  @override
+  Future<AndroidEqualizerBandSetGainResponse> androidEqualizerBandSetGain(
+      AndroidEqualizerBandSetGainRequest request) async {
+    return AndroidEqualizerBandSetGainResponse();
+  }
 }
 
 final byteRangeData = List.generate(200, (i) => i);
 
 class TestStreamAudioSource extends StreamAudioSource {
-  TestStreamAudioSource({dynamic tag}) : super(tag);
+  TestStreamAudioSource({dynamic tag}) : super(tag: tag);
 
   @override
   Future<StreamAudioResponse> request([int? start, int? end]) async {