+++ just_audio_platform_interface/lib/method_channel_just_audio.dart
@@ -6,7 +6,7 @@ import 'just_audio_platform_interface.dart';
 
 /// An implementation of [JustAudioPlatform] that uses method channels.
 class MethodChannelJustAudio extends JustAudioPlatform {
-  static final _mainChannel = MethodChannel('com.ryanheise.just_audio.methods');
+  static const _mainChannel = MethodChannel('com.ryanheise.just_audio.methods');
 
   @override
   Future<AudioPlayerPlatform> init(InitRequest request) async {
@@ -20,6 +20,14 @@ class MethodChannelJustAudio extends JustAudioPlatform {
     return DisposePlayerResponse.fromMap(
         (await _mainChannel.invokeMethod<Map<dynamic, dynamic>>(
             'disposePlayer', request.toMap()))!);
+  }
+
+  @override
+  Future<DisposeAllPlayersResponse> disposeAllPlayers(
+      DisposeAllPlayersRequest request) async {
+    return DisposeAllPlayersResponse.fromMap(
+        (await _mainChannel.invokeMethod<Map<dynamic, dynamic>>(
+            'disposeAllPlayers', request.toMap()))!);
   }
 }
 
@@ -37,6 +45,13 @@ class MethodChannelAudioPlayer extends AudioPlayerPlatform {
           .receiveBroadcastStream()
           .cast<Map<dynamic, dynamic>>()
           .map((map) => PlaybackEventMessage.fromMap(map));
+
+  @override
+  Stream<PlayerDataMessage> get playerDataMessageStream =>
+      EventChannel('com.ryanheise.just_audio.data.$id')
+          .receiveBroadcastStream()
+          .map((dynamic map) =>
+              PlayerDataMessage.fromMap(map as Map<dynamic, dynamic>));
 
   @override
   Future<LoadResponse> load(LoadRequest request) async {
@@ -69,6 +84,20 @@ class MethodChannelAudioPlayer extends AudioPlayerPlatform {
   }
 
   @override
+  Future<SetPitchResponse> setPitch(SetPitchRequest request) async {
+    return SetPitchResponse.fromMap((await _channel
+        .invokeMethod<Map<dynamic, dynamic>>('setPitch', request.toMap()))!);
+  }
+
+  @override
+  Future<SetSkipSilenceResponse> setSkipSilence(
+      SetSkipSilenceRequest request) async {
+    return SetSkipSilenceResponse.fromMap(
+        (await _channel.invokeMethod<Map<dynamic, dynamic>>(
+            'setSkipSilence', request.toMap()))!);
+  }
+
+  @override
   Future<SetLoopModeResponse> setLoopMode(SetLoopModeRequest request) async {
     return SetLoopModeResponse.fromMap((await _channel
         .invokeMethod<Map<dynamic, dynamic>>('setLoopMode', request.toMap()))!);
@@ -97,6 +126,25 @@ class MethodChannelAudioPlayer extends AudioPlayerPlatform {
     return SetAutomaticallyWaitsToMinimizeStallingResponse.fromMap(
         (await _channel.invokeMethod<Map<dynamic, dynamic>>(
             'setAutomaticallyWaitsToMinimizeStalling', request.toMap()))!);
+  }
+
+  @override
+  Future<SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse>
+      setCanUseNetworkResourcesForLiveStreamingWhilePaused(
+          SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest
+              request) async {
+    return SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse.fromMap(
+        (await _channel.invokeMethod<Map<dynamic, dynamic>>(
+            'setCanUseNetworkResourcesForLiveStreamingWhilePaused',
+            request.toMap()))!);
+  }
+
+  @override
+  Future<SetPreferredPeakBitRateResponse> setPreferredPeakBitRate(
+      SetPreferredPeakBitRateRequest request) async {
+    return SetPreferredPeakBitRateResponse.fromMap(
+        (await _channel.invokeMethod<Map<dynamic, dynamic>>(
+            'setPreferredPeakBitRate', request.toMap()))!);
   }
 
   @override
@@ -141,5 +189,38 @@ class MethodChannelAudioPlayer extends AudioPlayerPlatform {
     return ConcatenatingMoveResponse.fromMap(
         (await _channel.invokeMethod<Map<dynamic, dynamic>>(
             'concatenatingMove', request.toMap()))!);
+  }
+
+  @override
+  Future<AudioEffectSetEnabledResponse> audioEffectSetEnabled(
+      AudioEffectSetEnabledRequest request) async {
+    return AudioEffectSetEnabledResponse.fromMap(
+        (await _channel.invokeMethod<Map<dynamic, dynamic>>(
+            'audioEffectSetEnabled', request.toMap()))!);
+  }
+
+  @override
+  Future<AndroidLoudnessEnhancerSetTargetGainResponse>
+      androidLoudnessEnhancerSetTargetGain(
+          AndroidLoudnessEnhancerSetTargetGainRequest request) async {
+    return AndroidLoudnessEnhancerSetTargetGainResponse.fromMap(
+        (await _channel.invokeMethod<Map<dynamic, dynamic>>(
+            'androidLoudnessEnhancerSetTargetGain', request.toMap()))!);
+  }
+
+  @override
+  Future<AndroidEqualizerGetParametersResponse> androidEqualizerGetParameters(
+      AndroidEqualizerGetParametersRequest request) async {
+    return AndroidEqualizerGetParametersResponse.fromMap(
+        (await _channel.invokeMethod<Map<dynamic, dynamic>>(
+            'androidEqualizerGetParameters', request.toMap()))!);
+  }
+
+  @override
+  Future<AndroidEqualizerBandSetGainResponse> androidEqualizerBandSetGain(
+      AndroidEqualizerBandSetGainRequest request) async {
+    return AndroidEqualizerBandSetGainResponse.fromMap(
+        (await _channel.invokeMethod<Map<dynamic, dynamic>>(
+            'androidEqualizerBandSetGain', request.toMap()))!);
   }
 }