+++ just_audio_web/lib/just_audio_web.dart
@@ -35,14 +35,26 @@ class JustAudioPlugin extends JustAudioPlatform {
     players.remove(request.id);
     return DisposePlayerResponse();
   }
+
+  @override
+  Future<DisposeAllPlayersResponse> disposeAllPlayers(
+      DisposeAllPlayersRequest request) async {
+    for (var player in players.values) {
+      await player.release();
+    }
+    players.clear();
+    return DisposeAllPlayersResponse();
+  }
 }
 
 /// The web impluementation of [AudioPlayerPlatform].
 abstract class JustAudioPlayer extends AudioPlayerPlatform {
-  final _eventController = StreamController<PlaybackEventMessage>();
+  final _eventController = StreamController<PlaybackEventMessage>.broadcast();
+  final _dataEventController = StreamController<PlayerDataMessage>.broadcast();
   ProcessingStateMessage _processingState = ProcessingStateMessage.idle;
   bool _playing = false;
   int? _index;
+  double _speed = 1.0;
 
   /// Creates a platform player with the given [id].
   JustAudioPlayer({required String id}) : super(id);
@@ -50,6 +62,7 @@ abstract class JustAudioPlayer extends AudioPlayerPlatform {
   @mustCallSuper
   Future<void> release() async {
     _eventController.close();
+    _dataEventController.close();
   }
 
   /// Returns the current position of the player.
@@ -87,6 +100,7 @@ abstract class JustAudioPlayer extends AudioPlayerPlatform {
 /// An HTML5-specific implementation of [JustAudioPlayer].
 class Html5AudioPlayer extends JustAudioPlayer {
   final _audioElement = AudioElement();
+  late final _playPauseQueue = _PlayPauseQueue(_audioElement);
   Completer? _durationCompleter;
   AudioSourcePlayer? _audioSourcePlayer;
   LoopModeMessage _loopMode = LoopModeMessage.off;
@@ -119,6 +133,7 @@ class Html5AudioPlayer extends JustAudioPlayer {
       transition(ProcessingStateMessage.buffering);
     });
     _audioElement.addEventListener('canplaythrough', (event) {
+      _audioElement.playbackRate = _speed;
       transition(ProcessingStateMessage.ready);
     });
     _audioElement.addEventListener('progress', (event) {
@@ -175,6 +190,7 @@ class Html5AudioPlayer extends JustAudioPlayer {
             }
           }
         } else {
+          await _currentAudioSourcePlayer?.pause();
           transition(ProcessingStateMessage.completed);
         }
       }
@@ -195,11 +211,16 @@ class Html5AudioPlayer extends JustAudioPlayer {
       _eventController.stream;
 
   @override
+  Stream<PlayerDataMessage> get playerDataMessageStream =>
+      _dataEventController.stream;
+
+  @override
   Future<LoadResponse> load(LoadRequest request) async {
     _currentAudioSourcePlayer?.pause();
     _audioSourcePlayer = getAudioSource(request.audioSourceMessage);
     _index = request.initialIndex ?? 0;
-    final duration = await _currentAudioSourcePlayer!.load();
+    final duration = await _currentAudioSourcePlayer!
+        .load(request.initialPosition?.inMilliseconds);
     if (request.initialPosition != null) {
       await _currentAudioSourcePlayer!
           .seek(request.initialPosition!.inMilliseconds);
@@ -212,14 +233,19 @@ class Html5AudioPlayer extends JustAudioPlayer {
 
   /// Loads audio from [uri] and returns the duration of the loaded audio if
   /// known.
-  Future<Duration?> loadUri(final Uri uri) async {
+  Future<Duration?> loadUri(
+      final Uri uri, final Duration? initialPosition) async {
     transition(ProcessingStateMessage.loading);
     final src = uri.toString();
     if (src != _audioElement.src) {
       _durationCompleter = Completer<dynamic>();
       _audioElement.src = src;
+      _audioElement.playbackRate = _speed;
       _audioElement.preload = 'auto';
       _audioElement.load();
+      if (initialPosition != null) {
+        _audioElement.currentTime = initialPosition.inMilliseconds / 1000.0;
+      }
       try {
         await _durationCompleter!.future;
       } on MediaError catch (e) {
@@ -264,7 +290,7 @@ class Html5AudioPlayer extends JustAudioPlayer {
 
   @override
   Future<SetSpeedResponse> setSpeed(SetSpeedRequest request) async {
-    _audioElement.playbackRate = request.speed;
+    _audioElement.playbackRate = _speed = request.speed;
     return SetSpeedResponse();
   }
 
@@ -313,8 +339,7 @@ class Html5AudioPlayer extends JustAudioPlayer {
     if (index != _index) {
       _currentAudioSourcePlayer!.pause();
       _index = index;
-      await _currentAudioSourcePlayer!.load();
-      await _currentAudioSourcePlayer!.seek(position);
+      await _currentAudioSourcePlayer!.load(position);
       if (_playing) {
         _currentAudioSourcePlayer!.play();
       }
@@ -329,12 +354,14 @@ class Html5AudioPlayer extends JustAudioPlayer {
   @override
   Future<ConcatenatingInsertAllResponse> concatenatingInsertAll(
       ConcatenatingInsertAllRequest request) async {
+    final wasNotEmpty = _audioSourcePlayer?.sequence.isNotEmpty ?? false;
     _concatenating(request.id)!.setShuffleOrder(request.shuffleOrder);
     _concatenating(request.id)!
         .insertAll(request.index, getAudioSources(request.children));
-    if (_index != null && request.index <= _index!) {
+    if (_index != null && wasNotEmpty && request.index <= _index!) {
       _index = _index! + request.children.length;
     }
+    await _currentAudioSourcePlayer!.load();
     broadcastPlaybackEvent();
     return ConcatenatingInsertAllResponse();
   }
@@ -362,9 +389,11 @@ class Html5AudioPlayer extends JustAudioPlayer {
           _index = request.startIndex;
         }
         // Resume playback at the new item (if it exists)
-        if (_playing && _currentAudioSourcePlayer != null) {
+        if (_currentAudioSourcePlayer != null) {
           await _currentAudioSourcePlayer!.load();
-          _currentAudioSourcePlayer!.play();
+          if (_playing) {
+            _currentAudioSourcePlayer!.play();
+          }
         }
       } else if (request.endIndex <= _index!) {
         // Reflect that the current item has shifted its position
@@ -406,6 +435,20 @@ class Html5AudioPlayer extends JustAudioPlayer {
       setAutomaticallyWaitsToMinimizeStalling(
           SetAutomaticallyWaitsToMinimizeStallingRequest request) async {
     return SetAutomaticallyWaitsToMinimizeStallingResponse();
+  }
+
+  @override
+  Future<SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse>
+      setCanUseNetworkResourcesForLiveStreamingWhilePaused(
+          SetCanUseNetworkResourcesForLiveStreamingWhilePausedRequest
+              request) async {
+    return SetCanUseNetworkResourcesForLiveStreamingWhilePausedResponse();
+  }
+
+  @override
+  Future<SetPreferredPeakBitRateResponse> setPreferredPeakBitRate(
+      SetPreferredPeakBitRateRequest request) async {
+    return SetPreferredPeakBitRateResponse();
   }
 
   @override
@@ -501,7 +544,7 @@ abstract class IndexedAudioSourcePlayer extends AudioSourcePlayer {
       : super(html5AudioPlayer, id);
 
   /// Loads the audio for the underlying audio source.
-  Future<Duration?> load();
+  Future<Duration?> load([int? initialPosition]);
 
   /// Plays the underlying audio source.
   Future<void> play();
@@ -530,6 +573,8 @@ abstract class IndexedAudioSourcePlayer extends AudioSourcePlayer {
   /// The audio element that renders the audio.
   AudioElement get _audioElement => html5AudioPlayer._audioElement;
 
+  _PlayPauseQueue get _playPauseQueue => html5AudioPlayer._playPauseQueue;
+
   @override
   String toString() => "$runtimeType";
 }
@@ -544,6 +589,7 @@ abstract class UriAudioSourcePlayer extends IndexedAudioSourcePlayer {
   double? _resumePos;
   Duration? _duration;
   Completer? _completer;
+  int? _initialPos;
 
   UriAudioSourcePlayer(
       Html5AudioPlayer html5AudioPlayer, String id, this.uri, this.headers)
@@ -556,15 +602,22 @@ abstract class UriAudioSourcePlayer extends IndexedAudioSourcePlayer {
   List<int> get shuffleIndices => [0];
 
   @override
-  Future<Duration?> load() async {
-    _resumePos = 0.0;
-    return _duration = await html5AudioPlayer.loadUri(uri);
+  Future<Duration?> load([int? initialPosition]) async {
+    _initialPos = initialPosition;
+    _resumePos = (initialPosition ?? 0) / 1000.0;
+    _duration = await html5AudioPlayer.loadUri(
+        uri,
+        initialPosition != null
+            ? Duration(milliseconds: initialPosition)
+            : null);
+    _initialPos = null;
+    return _duration;
   }
 
   @override
   Future<void> play() async {
     _audioElement.currentTime = _resumePos!;
-    _audioElement.play();
+    await _playPauseQueue.play();
     _completer = Completer<dynamic>();
     await _completer!.future;
     _completer = null;
@@ -573,7 +626,7 @@ abstract class UriAudioSourcePlayer extends IndexedAudioSourcePlayer {
   @override
   Future<void> pause() async {
     _resumePos = _audioElement.currentTime as double?;
-    _audioElement.pause();
+    _playPauseQueue.pause();
     _interruptPlay();
   }
 
@@ -605,6 +658,7 @@ abstract class UriAudioSourcePlayer extends IndexedAudioSourcePlayer {
 
   @override
   Duration get position {
+    if (_initialPos != null) return Duration(milliseconds: _initialPos!);
     final seconds = _audioElement.currentTime as double;
     return Duration(milliseconds: (seconds * 1000).toInt());
   }
@@ -718,6 +772,7 @@ class ClippingAudioSourcePlayer extends IndexedAudioSourcePlayer {
   Completer<ClipInterruptReason>? _completer;
   double? _resumePos;
   Duration? _duration;
+  int? _initialPos;
 
   ClippingAudioSourcePlayer(Html5AudioPlayer html5AudioPlayer, String id,
       this.audioSourcePlayer, this.start, this.end)
@@ -729,16 +784,27 @@ class ClippingAudioSourcePlayer extends IndexedAudioSourcePlayer {
   @override
   List<int> get shuffleIndices => [0];
 
+  Duration get effectiveStart => start ?? Duration.zero;
+
   @override
-  Future<Duration?> load() async {
-    _resumePos = (start ?? Duration.zero).inMilliseconds / 1000.0;
-    final fullDuration =
-        (await html5AudioPlayer.loadUri(audioSourcePlayer.uri))!;
-    _audioElement.currentTime = _resumePos!;
-    _duration = Duration(
-        milliseconds: min((end ?? fullDuration).inMilliseconds,
-                fullDuration.inMilliseconds) -
-            (start ?? Duration.zero).inMilliseconds);
+  Future<Duration?> load([int? initialPosition]) async {
+    initialPosition ??= 0;
+    _initialPos = initialPosition;
+    final absoluteInitialPosition =
+        effectiveStart.inMilliseconds + initialPosition;
+    _resumePos = absoluteInitialPosition / 1000.0;
+    final fullDuration = (await html5AudioPlayer.loadUri(audioSourcePlayer.uri,
+        Duration(milliseconds: absoluteInitialPosition)));
+    _initialPos = null;
+    if (fullDuration != null) {
+      _duration = Duration(
+          milliseconds: min((end ?? fullDuration).inMilliseconds,
+                  fullDuration.inMilliseconds) -
+              effectiveStart.inMilliseconds);
+    } else if (end != null) {
+      _duration = Duration(
+          milliseconds: end!.inMilliseconds - effectiveStart.inMilliseconds);
+    }
     return _duration;
   }
 
@@ -747,10 +813,10 @@ class ClippingAudioSourcePlayer extends IndexedAudioSourcePlayer {
 
   @override
   Future<void> play() async {
-    _interruptPlay(ClipInterruptReason.simultaneous);
-    _audioElement.currentTime = _resumePos!;
-    _audioElement.play();
+    if (_completer != null) return;
     _completer = Completer<ClipInterruptReason>();
+    _audioElement.currentTime = _resumePos!;
+    await _playPauseQueue.play();
     ClipInterruptReason reason;
     while ((reason = await _completer!.future) == ClipInterruptReason.seek) {
       _completer = Completer<ClipInterruptReason>();
@@ -765,14 +831,14 @@ class ClippingAudioSourcePlayer extends IndexedAudioSourcePlayer {
   Future<void> pause() async {
     _interruptPlay(ClipInterruptReason.pause);
     _resumePos = _audioElement.currentTime as double?;
-    _audioElement.pause();
+    _playPauseQueue.pause();
   }
 
   @override
   Future<void> seek(int position) async {
     _interruptPlay(ClipInterruptReason.seek);
     _audioElement.currentTime =
-        _resumePos = start!.inMilliseconds / 1000.0 + position / 1000.0;
+        _resumePos = effectiveStart.inMilliseconds / 1000.0 + position / 1000.0;
   }
 
   @override
@@ -796,11 +862,10 @@ class ClippingAudioSourcePlayer extends IndexedAudioSourcePlayer {
 
   @override
   Duration get position {
+    if (_initialPos != null) return Duration(milliseconds: _initialPos!);
     final seconds = _audioElement.currentTime as double;
     var position = Duration(milliseconds: (seconds * 1000).toInt());
-    if (start != null) {
-      position -= start!;
-    }
+    position -= effectiveStart;
     if (position < Duration.zero) {
       position = Duration.zero;
     }
@@ -813,9 +878,7 @@ class ClippingAudioSourcePlayer extends IndexedAudioSourcePlayer {
       var seconds =
           _audioElement.buffered.end(_audioElement.buffered.length - 1);
       var position = Duration(milliseconds: (seconds * 1000).toInt());
-      if (start != null) {
-        position -= start!;
-      }
+      position -= effectiveStart;
       if (position < Duration.zero) {
         position = Duration.zero;
       }
@@ -836,7 +899,7 @@ class ClippingAudioSourcePlayer extends IndexedAudioSourcePlayer {
 }
 
 /// Reasons why playback of a clipping audio source may be interrupted.
-enum ClipInterruptReason { end, pause, seek, simultaneous }
+enum ClipInterruptReason { end, pause, seek }
 
 /// A player for a [LoopingAudioSourceMessage].
 class LoopingAudioSourcePlayer extends AudioSourcePlayer {
@@ -866,5 +929,44 @@ class LoopingAudioSourcePlayer extends AudioSourcePlayer {
       offset += childShuffleOrder.length;
     }
     return order;
+  }
+}
+
+class _PlayPauseRequest {
+  final bool playing;
+  final completer = Completer<void>();
+
+  _PlayPauseRequest(this.playing);
+}
+
+class _PlayPauseQueue {
+  final AudioElement audioElement;
+  final _queue = StreamController<_PlayPauseRequest>();
+
+  _PlayPauseQueue(this.audioElement) {
+    _run();
+  }
+
+  Future<void> play() async {
+    final request = _PlayPauseRequest(true);
+    _queue.add(request);
+    await request.completer.future;
+  }
+
+  Future<void> pause() async {
+    final request = _PlayPauseRequest(false);
+    _queue.add(request);
+    await request.completer.future;
+  }
+
+  Future<void> _run() async {
+    await for (var request in _queue.stream) {
+      if (request.playing) {
+        await audioElement.play();
+      } else {
+        audioElement.pause();
+      }
+      request.completer.complete();
+    }
   }
 }