+++ just_audio_platform_interface/CHANGELOG.md
@@ -1,3 +1,27 @@
+## 4.2.0
+
+- Add InitRequest.androidOffloadSchedulingEnabled.
+
+## 4.1.0
+
+- Add disposeAllPlayers method.
+
+## 4.0.0
+
+- Add playerDataMessageStream.
+- Add tag to IndexedAudioSourceMessage.
+
+## 3.1.0
+
+- Add setPitch.
+- Add setSkipSilence.
+- Add setCanUseNetworkResourcesForLiveStreamingWhilePaused.
+- Add setPreferredPeakBitRate.
+- Add audioEffectSetEnabled.
+- Add androidLoudnessEnhancerSetTargetGain.
+- Add androidEqualizerGetParameters.
+- Add androidEqualizerBandSetGain.
+
 ## 3.0.0
 
 - Null safety.