+++ just_audio_web/CHANGELOG.md
@@ -1,3 +1,43 @@
+## 0.4.7
+
+* Fix bug handling simultaneous play requests.
+
+## 0.4.6
+
+* Fix bug playing quickly after load.
+
+## 0.4.5
+
+* Fix interrupted play request bug.
+
+## 0.4.4
+
+* Implement disposeAllPlayers.
+
+## 0.4.3
+
+* Fix bug where setSpeed is forgotten after load().
+
+## 0.4.2
+
+* Fix bug loading audio after concat delete/insert.
+
+## 0.4.1
+
+* Remember position after stopping.
+
+## 0.4.0
+
+* Upgrade to platform interface 4.0.0
+
+## 0.3.2
+
+* Upgrade to platform interface 3.1.0
+
+## 0.3.1
+
+* Propagate play() exceptions (@twogood).
+
 ## 0.3.0
 
 * Null safety.